# java/loyalty-points-calculation/loyalty-project/LoyaltyPointsSwagger.yaml
swagger: '2.0'
info:
  description: Calculate Loyalty Points for Given User
  version: Version 1.0
  title: Loyalty Points API
  termsOfService: https://wso2.com/terms
  contact:
    name: Pamod Sylvester
    url: https://www.wso2.com
    email: pamod@wso2.com
  license:
    name: Apache 2.0
    url: https://wso2.com/license
host: localhost:8080
basePath: /
tags:
  - name: loyalty-point-controller
    description: Loyalty Point Controller
paths:
  /loyalty:
    get:
      tags:
        - loyalty-point-controller
      summary: getPoints
      operationId: getPointsUsingGET
      produces:
        - '*/*'
      parameters:
        - name: name
          in: query
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/LoyaltyPoint'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - loyalty-point-controller
      summary: addPoints
      operationId: addPointsUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: point
          description: point
          required: true
          schema:
            $ref: '#/definitions/LoyaltyPoint'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - loyalty-point-controller
      summary: deletePoints
      operationId: deletePointsUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: name
          in: query
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /loyalty/all:
    get:
      tags:
        - loyalty-point-controller
      summary: getAllPoints
      operationId: getAllPointsUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties:
              $ref: '#/definitions/LoyaltyPoint'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  LoyaltyPoint:
    type: object
    properties:
      nPoints:
        type: integer
        format: int32
      name:
        type: string
    title: LoyaltyPoint